+++ packages/logger/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @notjustcoders/one-logger-client-sdk
 
+## 1.0.2
+
+### Patch Changes
+
+- Introducing Class Wrapping Span Support
+
 ## 1.0.1
 
 ### Patch Changes